+++ src/index.ts
@@ -1,5 +1,6 @@
 import { RPReporter } from './reporter';
 export { ReportingApi } from './reportingApi';
 export { LOG_LEVELS, STATUSES } from './constants';
+export { RPTestInfo } from './constants';
 
 export default RPReporter;