+++ src/reporter.ts
@@ -37,9 +37,11 @@ import {
 } from './constants';
 import {
   calculateRpStatus,
+  getAdditionalInfo,
   getAgentInfo,
   getAttachments,
   getCodeRef,
+  getDescription,
   getSystemAttributes,
   isErrorLog,
   isFalse,
@@ -452,6 +454,9 @@ export class RPReporter implements Reporter {
     if (!savedTestItem) {
       return Promise.resolve();
     }
+
+    const additionalInfo = getAdditionalInfo(test);
+
     const {
       id: testItemId,
       attributes,
@@ -460,9 +465,12 @@ export class RPReporter implements Reporter {
       status: predefinedStatus,
     } = savedTestItem;
     let withoutIssue;
-    let testDescription = description;
+    let testDescription = getDescription(description, additionalInfo.description);
+
     const calculatedStatus = calculateRpStatus(test.outcome(), result.status, test.annotations);
-    const status = predefinedStatus || calculatedStatus;
+    const status = predefinedStatus || additionalInfo.status || calculatedStatus;
+    const calculatedTestCaseId = testCaseId || additionalInfo.testCaseId;
+    const mergedAttributes = additionalInfo.attributes.concat(attributes ?? []);
     if (status === STATUSES.SKIPPED) {
       withoutIssue = isFalse(this.config.skippedIssue);
     }
@@ -489,7 +497,9 @@ export class RPReporter implements Reporter {
         level: LOG_LEVELS.ERROR,
         message: stacktrace,
       });
-      testDescription = (description || '').concat(`\n\`\`\`error\n${stacktrace}\n\`\`\``);
+      testDescription = getDescription(description, additionalInfo.description).concat(
+        `\n\`\`\`error\n${stacktrace}\n\`\`\``,
+      );
     }
 
     [...this.nestedSteps.entries()].forEach(([key, value]) => {
@@ -511,9 +521,9 @@ export class RPReporter implements Reporter {
       endTime: this.client.helpers.now(),
       status,
       ...(withoutIssue && { issue: { issueType: 'NOT_ISSUE' } }),
-      ...(attributes && { attributes }),
+      ...(mergedAttributes.length !== 0 && { attributes: mergedAttributes }),
       ...(testDescription && { description: testDescription }),
-      ...(testCaseId && { testCaseId }),
+      ...(calculatedTestCaseId && { testCaseId: calculatedTestCaseId }),
     };
     const { promise } = this.client.finishTestItem(testItemId, finishTestItemObj);